+++ package.json
@@ -14,6 +14,7 @@
     "react-router-dom": "^6.8.1",
     "react-scripts": "^5.0.1",
     "react-slick": "^0.29.0",
+    "react-toastify": "^9.1.3",
     "slick-carousel": "^1.8.1",
     "web-vitals": "^2.1.4"
   },
@@ -45,7 +46,7 @@
   },
   "devDependencies": {
     "gh-pages": "^4.0.0",
-    "tailwindcss": "^3.2.4"
+    "tailwindcss": "^3.3.1"
   },
   "peerDependencies": {
     "react": ">=16.8.0"
