+++ src/App.css
@@ -7,12 +7,11 @@
   text-align: center;
 }
 
-
 .logo {
   font-family: "Romanesco", sans-serif;
   font-weight: 500;
   color: azure;
-  margin: -0.2% 0%;
+  margin: 0% 0%;
   padding: 9px 18px;
   border-radius: 50%;
   border: solid azure;
