+++ src/App.js
@@ -12,6 +12,7 @@ import { BlogsProvider } from "./components/getblogs";
 import { AuthProvider } from "./components/authContext";
 import Queries from "./components/admin/queries";
 import ProtectedRoute from "./protectedRoutes";
+
 function App() {
   return (
     <>
