+++ src/components/slider.jsx
@@ -6,6 +6,7 @@ import AIapp from "../assets/projects/AIapp.JPG";
 import musicapp from "../assets/projects/musicapp.JPG";
 import Capture from "../assets/projects/Capture.JPG";
 import reddy from "../assets/projects/reddy.png";
+import ecommerce from "../assets/projects/ecommerce.PNG";
 
 function SamplePrevArrow(props) {
   const { className, style, onClick } = props;
@@ -13,7 +14,7 @@ function SamplePrevArrow(props) {
     <div
       className={className}
       style={{
-        ...style
+        ...style,
       }}
       onClick={onClick}
     />
@@ -65,10 +66,22 @@ function SliderComponent() {
   };
   const slides = [
     {
-      img: AIapp,
-      name: "music app",
-      desc: "Welcome to our AI website, where we explore the latest developments in artificial intelligence and the groundbreaking work being done by OpenAI.",
-      demo: "https://github.com/Kevin-ishimwe/MusicWebApp",
+      img: Capture,
+      name: "health website",
+      desc: "I worked closely with the content team to ensure that the website's design and functionality aligned with the organization's mission to promote healthy living and well-being.",
+      demo: "https://agceramoz.com",
+    },
+    {
+      img: ecommerce,
+      name: "Ecommerce web app",
+      desc: "shop with ease from the comfort of your own home. Our platform is designed to provide you with a seamless online shopping experience, featuring a wide selection of products from various categories.",
+      demo: "https://ecommerce-app-legends-fe.vercel.app/",
+    },
+    {
+      img: musicapp,
+      name: "Music app",
+      desc: "Whether you're a casual listener, a serious musician, or somewhere in between, we've got you covered with the latest news, reviews, and insights from the world of music.",
+      demo: "https://jovial-profiterole-3bf938.netlify.app/",
     },
     {
       img: reddy,
@@ -77,16 +90,10 @@ function SliderComponent() {
       demo: "https://my-public-portfolio101.netlify.app/",
     },
     {
-      img: musicapp,
-      name: "AI app",
-      desc: "Whether you're a casual listener, a serious musician, or somewhere in between, we've got you covered with the latest news, reviews, and insights from the world of music.",
+      img: AIapp,
+      name: "OPEN AI API app",
+      desc: "Welcome to our AI website, where we explore the latest developments in artificial intelligence and the groundbreaking work being done by OpenAI.",
       demo: "https://github.com/Kevin-ishimwe/MusicWebApp",
-    },
-    {
-      img: Capture,
-      name: "health website",
-      desc: "I worked closely with the content team to ensure that the website's design and functionality aligned with the organization's mission to promote healthy living and well-being.",
-      demo: "https://agceramoz.com",
     },
   ];
   return (
