+++ src/index.css
@@ -69,6 +69,14 @@ code {
   width: 100%;
   height: 100%;
 }
+#div_see:active #div_see_text {
+  transition: 0.5s ease-in-out;
+  top: 0%;
+  padding: 50% 0%;
+  background: linear-gradient(0deg, rgb(0, 0, 0, 0.8), rgba(56, 56, 56, 0.389));
+  width: 100%;
+  height: 100%;
+}
 #dangerouslySetInnerHTML p {
   text-transform: none;
   font-family: Finlandica, serif;
